unsigned int problems. some could be bugs, others are just silly checks

+++ src/core/CLucene/search/PhraseQuery.cpp
@@ -303,7 +303,7 @@ CL_NS_DEF(search)
   //Post -
 
 	  //Get the length of terms
-      const size_t tpsLength = parentQuery->terms->size();
+      const int32_t tpsLength = (const int32_t)parentQuery->terms->size();
 
 	  //optimize zero-term case
       if (tpsLength == 0)
@@ -317,7 +317,7 @@ CL_NS_DEF(search)
     TermPositions* p = NULL;
 
 	//Iterate through all terms
-    for (size_t i = 0; i < tpsLength; i++) {
+    for (int32_t i = 0; i < tpsLength; i++) {
         //Get the termPostitions for the i-th term
         p = reader->termPositions((*parentQuery->terms)[i]);
